requirements.txt
streamlit
langchain
langchain-community
chromadb
pypdf2
python-dotenv
scikit-learn
